#1249 :truck: rename userStore admin tool
